+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## UNRELEASED 1.10.0
+
+- Dropped `scaffold-api` command
+- Dropped `generate-basic` command
+- Drop `httpx_client` from `APIClient`
+- Drop `httpx_async_client` from `APIClient`
+
 ## 1.9.2
 
 This release includes testing tools to make API integration testing easier.